Move sync to a separate module
